+++ crate_anon/crateweb/core/context_processors.py
@@ -33,7 +33,7 @@ from typing import Any, Dict
 from django.conf import settings
 from django.http.request import HttpRequest
 
-from crate_anon.common.constants import CRATE_DOCS_URL
+from crate_anon.common.constants import CRATE_DOCS_URL, HelpUrl
 
 
 # noinspection PyUnusedLocal
@@ -50,6 +50,7 @@ def common_context(request: HttpRequest) -> Dict[str, Any]:
     """
     return {
         'CRATE_DOCS_URL': CRATE_DOCS_URL,
+        'HelpUrl': HelpUrl,
         'nav_on_main_menu': False,
         'RESEARCH_DB_TITLE': settings.RESEARCH_DB_TITLE,
     }