+++ docs/source/nlp/run_gate_annie_demo_via_crate.sh
@@ -2,6 +2,8 @@
 
 CRATE_NLP_JAVA_CLASS_DIR=${CRATE_SOURCE_ROOT}/crate_anon/nlp_manager/compiled_nlp_classes
 
+# Note the unrealistic use of "STOP" as an end-of-input marker. Don't use that for real!
+
 java \
     -classpath "${CRATE_NLP_JAVA_CLASS_DIR}":"${GATE_DIR}/bin/gate.jar":"${GATE_DIR}/lib/*" \
     -Dgate.home="${GATE_DIR}" \